Renamed ChessServer\Command\Game\Async as AbstractSyncCommandChessServer\Command\Game\Sync
