+++ src/Command/Game/Sync/LeaveCommand.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace ChessServer\Command\Game\Async;
+namespace AbstractSyncCommandChessServer\Command\Game\Sync;
 
 use Chess\Variant\Classical\PGN\AN\Color;
 use ChessServer\Command\AbstractAsyncCommand;
-use ChessServer\Command\Game\Async\UpdateEloTask;
+use AbstractSyncCommandChessServer\Command\Game\Sync\UpdateEloTask;
 use ChessServer\Socket\AbstractSocket;
 
 class LeaveCommand extends AbstractAsyncCommand
