+++ src/Command/Game/Sync/RecognizerTask.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ChessServer\Command\Game\Async;
+namespace AbstractSyncCommandChessServer\Command\Game\Sync;
 
 use Chess\Media\ImgToPiecePlacement;
 use ChessServer\Command\AbstractSyncTask;
