+++ src/Command/Game/Sync/UpdateEloTask.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ChessServer\Command\Game\Async;
+namespace AbstractSyncCommandChessServer\Command\Game\Sync;
 
 use Chess\Elo\Game;
 use Chess\Elo\Player;
